refactor :  清除部分任务并且改变数据集结构
