+++ ark_nlp/dataset/biaffine_named_entity_recognition_dataset.py
@@ -16,10 +16,10 @@ import numpy as np
 import pandas as pd
 
 from torch.utils.data import Dataset
-from ark_nlp.dataset import BIOTokenClassificationDataset
+from ark_nlp.dataset import TokenClassificationDataset
 
 
-class BiaffineNERDataset(BIOTokenClassificationDataset):
+class BiaffineNERDataset(TokenClassificationDataset):
 
     def _get_categories(self):
         categories = sorted(list(set([label_['type'] for data in self.dataset for label_ in data['label']])))
