--- ark_nlp/model/ner/biaffine_ernie/__init__.py
@@ -1,17 +1,0 @@
-from ark_nlp.dataset import BiaffineNERDataset as Dataset 
-from ark_nlp.dataset import BiaffineNERDataset as BiaffineErnieNERDataset
-
-from ark_nlp.processor.tokenizer.transfomer import TokenTokenizer as Tokenizer
-from ark_nlp.processor.tokenizer.transfomer import TokenTokenizer as BiaffineErnieNERTokenizer
-
-from ark_nlp.nn import BertConfig as BiaffineErnieConfig
-from ark_nlp.model.ner.biaffine_Ernie.biaffine_Ernie import BiaffineErnie
-
-from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_model_optimizer
-from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_biaffine_Ernie_optimizer
-
-from ark_nlp.factory.task import BiaffineNERTask as Task
-from ark_nlp.factory.task import BiaffineNERTask as BiaffineErnieNERTask
-
-from ark_nlp.factory.predictor import BiaffineNERPredictor as Predictor
-from ark_nlp.factory.predictor import BiaffineNERPredictor as BiaffineErnieNERPredictor
